requirements.txt
aiogram==2.25.1
openai==1.3.5
python-dotenv==1.0.0
flake8==6.1.0
black==23.11.0 